+++ frontend/lib/app.dart
@@ -2,8 +2,9 @@
 
 import 'package:flutter/material.dart';
 import 'package:frontend/domain/utils/theme.dart';
+import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
 import 'package:frontend/presentation/screens/board/on_board.dart';
-
+import 'package:flutter_localizations/flutter_localizations.dart';
 
 
 class App extends StatefulWidget {
@@ -14,6 +15,13 @@ class App extends StatefulWidget {
 }
 
 class _AppState extends State<App> {
+  Locale? _locale;
+
+  void _changeLanguage(Locale locale) {
+    setState(() {
+      _locale = locale;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +32,15 @@ class _AppState extends State<App> {
         theme: AppTheme.lightTheme, // Apply the light theme
         darkTheme: AppTheme.darkTheme, // Apply the dark theme (optional)
         title: 'HeartBeats',
-        home: OnBoardScreen()
+         locale: _locale,
+      supportedLocales: S.delegate.supportedLocales,
+      localizationsDelegates: [
+        S.delegate,
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+        home: OnBoardScreen(onLocaleChange: _changeLanguage)
       
     );
   }