+++ frontend/lib/presentation/screens/board/small_widget.dart
@@ -3,18 +3,29 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:frontend/domain/utils/paths.dart';
+import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
+import 'package:frontend/presentation/boilerplate/app_bar.dart';
 
 
 class SmallWidget extends StatefulWidget {
-  const SmallWidget({super.key});
+  final Function(Locale) onLocaleChange;
+
+  const SmallWidget({
+    super.key, required this.onLocaleChange
+    });
 
   @override
   SmallWidgetState createState() => SmallWidgetState();
 }
 
 class SmallWidgetState extends State<SmallWidget> {
+
+  
+
   @override
   Widget build(BuildContext context) {
+
+    
     
     return Stack(
         children: [
@@ -35,33 +46,39 @@ class SmallWidgetState extends State<SmallWidget> {
           ), Scaffold(
              extendBodyBehindAppBar: true, // Allows content to extend behind the AppBar
             backgroundColor: Colors.transparent,
-        appBar: AppBar(
-          backgroundColor: Colors.transparent,
-          leading: Align(
-            alignment: Alignment.center, // Ensures the logo is aligned to the left
-            child: SvgPicture.asset(
-            '${DirPath.media}logo.svg', // Path to your logo
-            height: 40, // Adjust size
-          ),
-
-           // Optional: Center the logo
-        ),
-      
-        ),
+        appBar: MyAppBar(onLocaleChange: widget.onLocaleChange),
 
         body: SingleChildScrollView(
           child: Column( 
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            crossAxisAlignment: CrossAxisAlignment.center,
             children: [ 
               Center(
           child: SvgPicture.asset(
             '${DirPath.media}on_board_background.svg', // Path to your logo
             height: 500, // Adjust size
           )),
-          
-         Text(
-          "text for test",
-          style: const TextStyle(fontSize: 24, color: Colors.black),
-        ),
+          SizedBox(height: 70),
+         ElevatedButton.icon(
+            onPressed: () {
+              // Define what happens when the button is pressed
+            },
+            icon: const Icon(Icons.music_note_sharp), // Your icon
+            label: Text(
+               S.of(context).listenMusic,
+               style: TextStyle(color: Theme.of(context).colorScheme.onSecondary),
+            ),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Theme.of(context).colorScheme.onSurface,
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(5), // Rounded corners
+              ),
+              padding: const EdgeInsets.symmetric(
+                horizontal: 20,
+                vertical: 20,
+              ), // Padding inside the button
+            ),
+          ),
       
           
           ])),